backend/requirements.txt
appnope==0.1.2
asgiref==3.4.1
autopep8==1.6.0
Babel==2.9.1
backcall==0.2.0
backports.entry-points-selectable==1.1.0
black==21.9b0
boto3==1.21.27
botocore==1.24.27
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.3
cryptography==36.0.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.7
django-allauth==0.46.0
django-cors-headers==3.11.0
django-dotenv==1.4.2
django-filter==21.1
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
entrypoints==0.3
filelock==3.0.12
future==0.18.2
GDAL==3.4.2
geocoder==1.38.1
gunicorn==20.1.0
idna==3.3
ipykernel==6.6.0
ipython==7.30.1
jedi==0.18.1
jmespath==1.0.0
jupyter-client==7.1.0
jupyter-core==4.9.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
nest-asyncio==1.5.4
numpy==1.22.3
oauthlib==3.1.1
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
pipenv==2021.11.23
platformdirs==2.3.0
prompt-toolkit==3.0.24
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.11.1
PyJWT==2.3.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
pyzmq==22.3.0
ratelim==0.1.6
regex==2021.10.23
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.2
tornado==6.1
traitlets==5.1.1
typing-extensions==3.10.0.2
urllib3==1.26.7
virtualenv==20.7.2
virtualenv-clone==0.5.7
wcwidth==0.2.5
whitenoise==6.0.0
